dashboard/requirements.txt
streamlit
pandas
plotly
awswrangler
python-dotenv